[None][fix] Refactor KV cache V2 pending stats
