+++ tensorrt_llm/runtime/kv_cache_manager_v2/setup_mypyc.py
@@ -86,6 +86,7 @@ modules = [
     "kv_cache_manager_v2/_core/__init__.py",
     "kv_cache_manager_v2/_core/_kv_cache_manager.py",
     "kv_cache_manager_v2/_core/_kv_cache.py",
+    "kv_cache_manager_v2/_core/_pending_stats.py",
     # _eviction_controller submodule
     "kv_cache_manager_v2/_eviction_controller/__init__.py",
     "kv_cache_manager_v2/_eviction_controller/_eviction_controller.py",
